Split GenerateReleaseArtifacts into three tasks
This change splits the GenerateReleaseArtifacts into three tasks
and moves them to a separate project (and nuget package). The three
new tasks are: CreateChangeLogEntry, CreateReleaseNotesHtml, and
StampChangelogFileWithVersion.

+semver: breaking
