+++ Readme.md
@@ -2,6 +2,16 @@
 
 Several useful msbuild tasks.
 
+## Current Tasks
+
+### [`SIL.ReleaseTasks` package](Documentation/SIL.ReleaseTasks.md)
+
+**Task**                      | **Description**
+------------------------------|----------------------------------------------------------------
+[CreateChangelogEntry](Documentation/SIL.ReleaseTasks.md#createchangelogentry-task) | Given a Changelog file, this task will add an entry to the debian changelog.
+[CreateReleaseNotesHtml](Documentation/SIL.ReleaseTasks.md#createreleasenoteshtml-task) | Given a markdown-style changelog file, this class will generate a release notes HTML file.
+[StampChangelogFileWithVersion](Documentation/SIL.ReleaseTasks.md#stampchangelogfilewithversion-task) | Replaces the first line in a markdown-style Changelog/Release file with the version and date.
+
 ### [`SIL.BuildTasks` package](Documentation/SIL.BuildTasks.md)
 
 **Task**                      | **Description**
