+++ SIL.ReleaseTasks.Tests/SIL.ReleaseTasks.Tests.csproj
@@ -1,0 +1,27 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFrameworks>net461</TargetFrameworks>
+    <RootNamespace>SIL.ReleaseTasks.Tests</RootNamespace>
+    <Configurations>Debug;Release</Configurations>
+    <OutputPath>../output/$(Configuration)</OutputPath>
+    <AssemblyTitle>SIL.ReleaseTasks.Tests</AssemblyTitle>
+    <Company>SIL</Company>
+    <Product>Palaso Library</Product>
+    <Copyright>Copyright © 2018 SIL International</Copyright>
+    <ErrorReport>prompt</ErrorReport>
+    <WarningLevel>4</WarningLevel>
+  </PropertyGroup>
+  <ItemGroup>
+    <Reference Include="Microsoft.Build" />
+    <Reference Include="Microsoft.Build.Framework" />
+    <Reference Include="Microsoft.Build.Utilities.v4.0" />
+  </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.7.0" />
+    <PackageReference Include="NUnit" Version="3.10.1" />
+    <PackageReference Include="NUnit3TestAdapter" Version="3.10.0" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\SIL.ReleaseTasks\SIL.ReleaseTasks.csproj" />
+  </ItemGroup>
+</Project>
